Fixing test suite.
